Add controlled particle boundary limits...
Note that this commit doesn't yet mark a fully-functional bounding feature,
since when a few blue particles are located along an edge and are absorbed, the
moving particle can get stuck there (for quite obvious reasons that I don't
mind typing).
The feature works satisfyingly well, but should have this BUG fixed in the near
future.